✅ Fix tests for transactions api utils
